+++ src/main/java/ml/pluto7073/pdapi/PDAPI.java
@@ -15,10 +15,13 @@ import ml.pluto7073.pdapi.specialty.SpecialtyDrinkManager;
 import ml.pluto7073.pdapi.util.DrinkUtil;
 import ml.pluto7073.plutonium.PlutoniumConfig;
 import ml.pluto7073.plutonium.config.ServerConfigType;
+import net.fabricmc.api.EnvType;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.client.Minecraft;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
@@ -39,6 +42,8 @@ public class PDAPI implements ModInitializer {
     public static final ResourceKey<CreativeModeTab> SPECIALTY_DRINKS_TAB = ResourceKey.create(Registries.CREATIVE_MODE_TAB, asId("specialty_drinks"));
     public static final ServerConfigType<PDCommonConfig> CONFIG_TYPE =
             Registry.register(PlutoniumConfig.SERVER_CONFIG_TYPES, asId("common"), new ServerConfigType<>(PDCommonConfig.INSTANCE, PDCommonConfig::new));
+    public static final DrinkAdditionManager SERVER_ADDITION_MANAGER = new DrinkAdditionManager();
+    public static final SpecialtyDrinkManager SERVER_SPECIALITY_DRINK_MANAGER = new SpecialtyDrinkManager();
 
 
     @Override
@@ -51,30 +56,19 @@ public class PDAPI implements ModInitializer {
         PDItems.init();
         PDMobEffects.init();
 
-        ResourceManagerHelper.get(PackType.SERVER_DATA).registerReloadListener(new DrinkAdditionManager());
-        ResourceManagerHelper.get(PackType.SERVER_DATA).registerReloadListener(new SpecialtyDrinkManager());
-
-        DrinkUtil.registerOldToNewConverter("Coffee/Additions", tag -> {
-            if (!(tag instanceof ListTag list)) return tag;
-            if (list.isEmpty()) return list;
-            for (int i = 0; i < list.size(); i++) {
-                ResourceLocation id = new ResourceLocation(list.getString(i));
-                boolean ogCoffee = !DrinkAdditionManager.containsId(id) && DrinkAdditionManager.containsId(PDAPI.asId(id.getPath()));
-                if (!ogCoffee) continue;
-                id = PDAPI.asId(id.getPath());
-                list.set(i, DrinkUtil.stringAsNbt(id.toString()));
-            }
-            return list;
-        });
+        ResourceManagerHelper.get(PackType.SERVER_DATA).registerReloadListener(SERVER_ADDITION_MANAGER);
+        ResourceManagerHelper.get(PackType.SERVER_DATA).registerReloadListener(SERVER_SPECIALITY_DRINK_MANAGER);
 
         PDScreens.init();
 
         Registry.register(BuiltInRegistries.CREATIVE_MODE_TAB, SPECIALTY_DRINKS_TAB, FabricItemGroup.builder().icon(() -> new ItemStack(PDItems.ICON))
                 .title(Component.translatable("creative_tab.pdapi.specialty_drinks")).build());
         ItemGroupEvents.modifyEntriesEvent(SPECIALTY_DRINKS_TAB).register(stacks -> {
-            for (SpecialtyDrink d : SpecialtyDrinkManager.values()
-                    .stream().sorted(DrinkUtil.alphabetizer(SpecialtyDrink::languageKey)).toList()) {
-                stacks.accept(d.getAsItem());
+            if (FabricLoader.getInstance().getEnvironmentType() != EnvType.CLIENT) return;
+            if (Minecraft.getInstance().level == null) return;
+            for (SpecialtyDrink d : Minecraft.getInstance().level.getSpecialtyDrinkManager().values()
+                    .stream().sorted(DrinkUtil.alphabetizer(drink -> drink.languageKey(Minecraft.getInstance().level))).toList()) {
+                stacks.accept(d.getAsItem(Minecraft.getInstance().level));
             }
         });