+++ src/main/java/ml/pluto7073/pdapi/addition/DrinkAdditionManager.java
@@ -1,14 +1,9 @@
 package ml.pluto7073.pdapi.addition;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
 import com.mojang.serialization.JsonOps;
-import ml.pluto7073.chemicals.Chemicals;
 import ml.pluto7073.pdapi.PDAPI;
-import ml.pluto7073.pdapi.PDRegistries;
-import ml.pluto7073.pdapi.addition.action.OnDrinkAction;
-import ml.pluto7073.pdapi.addition.action.OnDrinkSerializer;
 import ml.pluto7073.pdapi.networking.packet.clientbound.ClientboundSyncAdditionRegistryPacket;
 import ml.pluto7073.pdapi.util.DrinkUtil;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
@@ -28,32 +23,25 @@ import java.util.*;
 
 public class DrinkAdditionManager implements SimpleSynchronousResourceReloadListener {
 
-    private static final Map<ResourceLocation, DrinkAddition> REGISTRY = new HashMap<>();
-    private static final Map<ResourceLocation, DrinkAddition> STATIC_REGISTRY = new HashMap<>();
-
     public static final String ADDITIONS_NBT_KEY = "Additions";
-    public static final DrinkAddition EMPTY = register(PDAPI.asId("empty"), new DrinkAddition.Builder().build());
     public static final ResourceLocation PHASE = PDAPI.asId("phase/additions");
+
+    private final Map<ResourceLocation, DrinkAddition> registry = new HashMap<>();
 
     public DrinkAdditionManager() {
         ServerLifecycleEvents.SYNC_DATA_PACK_CONTENTS.register(PHASE, (player, joined) -> send(player));
     }
 
-    public static DrinkAddition register(ResourceLocation id, DrinkAddition addition) {
-        return register(id, addition, true);
-    }
-
-    public static DrinkAddition register(ResourceLocation id, DrinkAddition addition, boolean staticAdd) {
+    public DrinkAddition register(ResourceLocation id, DrinkAddition addition) {
         if (containsId(id)) {
             if (get(id).getCurrentWeight() >= addition.getCurrentWeight()) return get(id);
         }
-        REGISTRY.put(id, addition);
-        if (staticAdd) STATIC_REGISTRY.put(id, addition);
+        registry.put(id, addition);
         return addition;
     }
 
-    public static ResourceLocation getId(DrinkAddition addition) {
-        for (Map.Entry<ResourceLocation, DrinkAddition> entry : REGISTRY.entrySet()) {
+    public ResourceLocation getId(DrinkAddition addition) {
+        for (Map.Entry<ResourceLocation, DrinkAddition> entry : registry.entrySet()) {
             if (Objects.equals(entry.getValue(), addition)) {
                 return entry.getKey();
             }
@@ -61,34 +49,33 @@ public class DrinkAdditionManager implements SimpleSynchronousResourceReloadList
         throw new IllegalArgumentException("Unregistered drink addition: " + addition.toString());
     }
 
-    public static DrinkAddition get(ResourceLocation id) {
-        return REGISTRY.get(id);
+    public DrinkAddition get(ResourceLocation id) {
+        return registry.get(id);
     }
 
-    public static void resetRegistry() {
-        REGISTRY.clear();
-        REGISTRY.putAll(STATIC_REGISTRY);
+    public void resetRegistry() {
+        registry.clear();
     }
 
-    public static boolean containsId(ResourceLocation id) {
-        return REGISTRY.containsKey(id);
+    public boolean containsId(ResourceLocation id) {
+        return registry.containsKey(id);
     }
 
-    public static boolean containsAddition(DrinkAddition addition) {
-        return REGISTRY.containsValue(addition);
+    public boolean containsAddition(DrinkAddition addition) {
+        return registry.containsValue(addition);
     }
 
-    public static boolean contains(ResourceLocation id, DrinkAddition addition) {
+    public boolean contains(ResourceLocation id, DrinkAddition addition) {
         return containsId(id) && containsAddition(addition) && get(id).equals(addition);
     }
 
-    public static boolean contains(Map.Entry<ResourceLocation, DrinkAddition> entry) {
+    public boolean contains(Map.Entry<ResourceLocation, DrinkAddition> entry) {
         return contains(entry.getKey(), entry.getValue());
     }
 
-    public static void send(ServerPlayer entity) {
+    public void send(ServerPlayer entity) {
 
-        ServerPlayNetworking.send(entity, new ClientboundSyncAdditionRegistryPacket(REGISTRY));
+        ServerPlayNetworking.send(entity, new ClientboundSyncAdditionRegistryPacket(registry));
 
     }
 
@@ -118,10 +105,10 @@ public class DrinkAdditionManager implements SimpleSynchronousResourceReloadList
                 }
 
                 register(id, DrinkAddition.CODEC.parse(JsonOps.INSTANCE, object).getOrThrow(false, s -> {
-                    throw new IllegalStateException(s);
-                }), false);
+                    throw new JsonParseException(s);
+                }));
                 i++;
-            } catch (IOException e) {
+            } catch (Exception e) {
                 PDAPI.LOGGER.error("Could not load Drink Addition {}", id, e);
             }
         }