+++ src/main/java/ml/pluto7073/pdapi/item/AbstractCustomizableDrinkItem.java
@@ -8,6 +8,7 @@ import ml.pluto7073.pdapi.addition.DrinkAdditionManager;
 import net.minecraft.ChatFormatting;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
@@ -44,18 +45,18 @@ public abstract class AbstractCustomizableDrinkItem extends Item implements Chem
         this.baseItem = baseItem;
     }
 
-    public double getTotalVolume(ItemStack stack) {
+    public double getTotalVolume(ItemStack stack, Level level) {
         double vol = baseVolume;
-        for (DrinkAddition a : DrinkUtil.getAdditionsFromStack(stack)) {
+        for (DrinkAddition a : DrinkUtil.getAdditionsFromStack(stack, level)) {
             vol += a.volume();
         }
         return vol;
     }
 
     @Override
-    public float getChemicalContent(ResourceLocation name, ItemStack stack) {
+    public float getChemicalContent(ResourceLocation name, ItemStack stack, Level level) {
         int amount = 0;
-        for (DrinkAddition a : DrinkUtil.getAdditionsFromStack(stack)) {
+        for (DrinkAddition a : DrinkUtil.getAdditionsFromStack(stack, level)) {
             if (a.getChemicals().containsKey(name))
                 amount += a.getChemicals().get(name);
         }
@@ -66,9 +67,11 @@ public abstract class AbstractCustomizableDrinkItem extends Item implements Chem
     public float getConsumedChemicalContent(ResourceLocation id, ItemStack stack) {
         float amount = getChemicalContent(id, stack);
         return (float) ((getSipAmount(stack) / 2f) / getTotalVolume(stack)) * amount;
+    public float getConsumedChemicalContent(ResourceLocation id, ItemStack stack, Level level) {
+        float amount = getChemicalContent(id, stack, level);
     }
 
-    protected Item baseItem(ItemStack stack) {
+    protected Item baseItem(ItemStack stack, Level level) {
         return baseItem;
     }
 
@@ -87,8 +90,11 @@ public abstract class AbstractCustomizableDrinkItem extends Item implements Chem
         return ItemUtils.startUsingInstantly(world, user, hand);
     }
 
-    protected int getSipAmount(ItemStack stack) {
-        return getTotalVolume(stack) >= 10 ? 2 : 1;
+    /**
+     * @return the amount of liquid to be sipped.  1oz for >10oz drinks, 0.5oz for <10oz drinks
+     */
+    protected double getSipAmount(ItemStack stack, Level level) {
+        return getTotalVolume(stack, level) > 10 ? 1 : 0.5;
     }
 
     @Override
@@ -106,7 +112,7 @@ public abstract class AbstractCustomizableDrinkItem extends Item implements Chem
         }
 
         if (!world.isClientSide) {
-            DrinkAddition[] additions = DrinkUtil.getAdditionsFromStack(stack);
+            DrinkAddition[] additions = DrinkUtil.getAdditionsFromStack(stack, world);
             for (DrinkAddition addition : additions) {
                 addition.onDrink(stack, world, user);
             }
@@ -121,11 +127,11 @@ public abstract class AbstractCustomizableDrinkItem extends Item implements Chem
 
         if (player == null || !player.getAbilities().instabuild) {
             if (stack.isEmpty()) {
-                return new ItemStack(baseItem(stack));
+                return new ItemStack(baseItem(stack, world));
             }
 
             if (player != null) {
-                player.getInventory().add(new ItemStack(baseItem(stack)));
+                player.getInventory().add(new ItemStack(baseItem(stack, world)));
             }
         }
         user.gameEvent(GameEvent.DRINK);
@@ -135,16 +141,18 @@ public abstract class AbstractCustomizableDrinkItem extends Item implements Chem
     @Override
     public void appendHoverText(ItemStack stack, @Nullable Level world, List<Component> tooltip, TooltipFlag context) {
         DrinkAddition[] addIns = DrinkUtil.getAdditionsFromStack(stack);
+        if (world == null) return;
+        DrinkAddition[] addIns = DrinkUtil.getAdditionsFromStack(stack, world);
         HashMap<ResourceLocation, Integer> additionCounts = new HashMap<>();
         for (DrinkAddition addIn : addIns) {
-            if (addIn == DrinkAdditionManager.EMPTY || addIn == null) continue;
-            ResourceLocation id = DrinkAdditionManager.getId(addIn);
+            if (addIn == null) continue;
+            ResourceLocation id = world.getDrinkAdditionManager().getId(addIn);
             if (additionCounts.containsKey(id)) {
                 int count = additionCounts.get(id);
                 additionCounts.put(id, ++count);
             } else additionCounts.put(id, 1);
         }
-        additionCounts.forEach((id, count) -> tooltip.add(Component.translatable(DrinkAdditionManager.get(id).getTranslationKey(), count).withStyle(ChatFormatting.GRAY)));
+        additionCounts.forEach((id, count) -> tooltip.add(Component.translatable(world.getDrinkAdditionManager().get(id).getTranslationKey(world), count).withStyle(ChatFormatting.GRAY)));
     }
 
     @Override