+++ src/main/java/ml/pluto7073/pdapi/item/SpecialtyDrinkItem.java
@@ -3,7 +3,11 @@ package ml.pluto7073.pdapi.item;
 import ml.pluto7073.pdapi.PDAPI;
 import ml.pluto7073.pdapi.util.DrinkUtil;
 import ml.pluto7073.pdapi.specialty.SpecialtyDrink;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.MethodsReturnNonnullByDefault;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
@@ -19,25 +23,25 @@ public class SpecialtyDrinkItem extends AbstractCustomizableDrinkItem {
     }
 
     @Override
-    protected Item baseItem(ItemStack stack) {
-        SpecialtyDrink drink = DrinkUtil.getSpecialDrink(stack);
+    protected Item baseItem(ItemStack stack, Level level) {
+        SpecialtyDrink drink = DrinkUtil.getSpecialDrink(stack, level);
         AbstractCustomizableDrinkItem base =
                 (AbstractCustomizableDrinkItem) drink.getBaseItem(stack).getItem();
-        return base.baseItem(stack);
+        return base.baseItem(stack, level);
     }
 
     @Override
-    public double getTotalVolume(ItemStack stack) {
+    public double getTotalVolume(ItemStack stack, Level level) {
         try {
-            SpecialtyDrink specialty = DrinkUtil.getSpecialDrink(stack);
+            SpecialtyDrink specialty = DrinkUtil.getSpecialDrink(stack, level);
             if (specialty.volume() != 0) {
                 return specialty.volume();
             }
             ItemStack baseItem = specialty.getBaseItem(stack);
             if (baseItem.getItem() instanceof AbstractCustomizableDrinkItem drink) {
-                return drink.getTotalVolume(baseItem);
+                return drink.getTotalVolume(baseItem, level);
             } else {
-                return super.getTotalVolume(stack);
+                return super.getTotalVolume(stack, level);
             }
         } catch (Exception e) {
             PDAPI.LOGGER.warn("Error getting total volume of {}", stack, e);
@@ -46,16 +50,16 @@ public class SpecialtyDrinkItem extends AbstractCustomizableDrinkItem {
     }
 
     @Override
-    public float getChemicalContent(ResourceLocation name, ItemStack stack) {
+    public float getChemicalContent(ResourceLocation name, ItemStack stack, Level level) {
         float amount;
         try {
-            SpecialtyDrink specialty = DrinkUtil.getSpecialDrink(stack);
+            SpecialtyDrink specialty = DrinkUtil.getSpecialDrink(stack, level);
             amount = specialty.chemicals().getOrDefault(name, 0f);
             ItemStack baseItem = specialty.getBaseItem(stack);
             if (baseItem.getItem() instanceof AbstractCustomizableDrinkItem drink) {
-                return amount + drink.getChemicalContent(name, baseItem);
+                return amount + drink.getChemicalContent(name, baseItem, level);
             } else {
-                return super.getChemicalContent(name, stack) + amount;
+                return super.getChemicalContent(name, stack, level) + amount;
             }
         } catch (Exception e) {
             PDAPI.LOGGER.warn("Error getting amount of {} in {}", name, stack, e);
@@ -65,9 +69,9 @@ public class SpecialtyDrinkItem extends AbstractCustomizableDrinkItem {
 
     @Override
     public ItemStack finishUsingItem(ItemStack stack, Level level, LivingEntity user) {
-        SpecialtyDrink drink = DrinkUtil.getSpecialDrink(stack);
+        SpecialtyDrink drink = DrinkUtil.getSpecialDrink(stack, level);
 
-        if (!level.isClientSide) drink.actions().forEach(action -> action.onDrink(stack, level, user));
+        if (!level.isClientSide) drink.actions(level).forEach(action -> action.onDrink(stack, level, user));
 
         return super.finishUsingItem(stack, level, user);
     }
@@ -75,8 +79,12 @@ public class SpecialtyDrinkItem extends AbstractCustomizableDrinkItem {
     @Override
     public String getDescriptionId(ItemStack stack) {
         try {
-            SpecialtyDrink drink = DrinkUtil.getSpecialDrink(stack);
-            return drink.name();
+            if (FabricLoader.getInstance().getEnvironmentType() == EnvType.CLIENT && Minecraft.getInstance().level != null) {
+                Level level = Minecraft.getInstance().level;
+                return DrinkUtil.getSpecialDrink(stack, level).name(level);
+            } else {
+                return new ResourceLocation(stack.getOrCreateTag().getString("Drink")).toLanguageKey("drink");
+            }
         } catch (Exception e) {
             return super.getDescriptionId(stack);
         }